pyproject.toml
[project]
name = "crewai-application-tracking-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=1.7.0",
    "crewai-tools>=1.7.0",
    "fastapi>=0.124.2",
    "ipykernel>=7.1.0",
    "langchain-groq>=1.1.0",
    "litellm>=1.75.3",
    "pydantic>=2.11.10",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
